Enable Libdl in PCRE
