+++ base/Base.jl
@@ -106,6 +106,27 @@ include("lock.jl")
 # strings & printing
 include("intfuncs.jl")
 include("strings/strings.jl")
+
+#=
+isdebugbuild is defined here as this is imported in libdl.jl (included in libc.jl)
+=#
+"""
+    isdebugbuild()
+
+Return `true` if julia is a debug version.
+"""
+function isdebugbuild()
+    return ccall(:jl_is_debugbuild, Cint, ()) != 0
+end
+
+# Enable dynamic library loading
+include("osinfo.jl") # Defines the module Sys, more added later
+include("path.jl") # Defines the module Filesystem, more content added later
+using .Filesystem 
+include("libc.jl") # Libdl (include in libc.jl) is required for regex.jl
+using .Libc: getpid, gethostname, time, memcpy, memset, memmove, memcmp
+
+# More strings & printing
 include("regex.jl")
 include("parse.jl")
 include("shell.jl")
@@ -130,16 +151,8 @@ include("missing.jl")
 # version
 include("version.jl")
 
-#=
-isdebugbuild is defined here as this is imported in libdl.jl (included in libc.jl)
-The method is added in util.jl
-=#
-function isdebugbuild end
-
 # system & environment
-include("sysinfo.jl")
-include("libc.jl")
-using .Libc: getpid, gethostname, time, memcpy, memset, memmove, memcmp
+Core.eval(Sys, :(include("sysinfo.jl")))
 
 const USING_STOCK_GC = occursin("stock", GC.gc_active_impl())
 
@@ -175,8 +188,7 @@ include("libuv.jl")
 include("asyncevent.jl")
 include("iostream.jl")
 include("stream.jl")
-include("filesystem.jl")
-using .Filesystem
+Core.eval(Filesystem, :(include("filesystem.jl")))
 include("cmd.jl")
 include("process.jl")
 include("terminfo.jl")
